adding cart logic, parse logic and taxcalculator with tests

+++ src/main/java/com/andersonfelipe/taxcalculator/tax/Tax.java
@@ -6,12 +6,12 @@ public abstract class Tax {
 	public abstract boolean isApplicable(Product product);
 	abstract float getRate();
 	
-	public double calculate(Product product) {
+	public float calculate(Product product) {
 		if(isApplicable(product)) {
 			
-			double tax = (product.getShelfPrice() * getRate()) / 100;
+			float tax = (product.getShelfPrice() * getRate()) / 100;
 			
-			tax = Math.ceil(tax / 0.05) * 0.05;
+			tax = (float) (Math.ceil(tax / 0.05) * 0.05);
 			
 			return tax;
 		}
